requirements.txt
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
Flask>=2.0
python-dotenv>=0.19
firebase-admin>=6.0
gunicorn>=21.0
